Switch Prompt to ';'
For a long time I used '%' as my prompt. Previously, I used UTF8 glyphs
but those are obviously only fancy if the terminal supports them as
well. So, I switched to '%'.

Anyway, for quite some time I was thinking of either using '#' to be
able to copy / paste the lines with the commands on it "somewhere" and
technically they would be ok - except that they don't trigger anything
coz they would be comments.

Lately I was reading for other reasons @thingskatedid … and decided to
give https://twitter.com/thingskatedid/status/1316081075043463170 a
chance.